move filter to the components directory for easier reuse
