+++ components/gallery/GalleryItem.jsx
